+++ source/icu.net.tests/UnicodeStringTests.cs
@@ -1,6 +1,7 @@
 // Copyright (c) 2013 SIL International
 // This software is licensed under the MIT license (http://opensource.org/licenses/MIT)
 using System;
+using Icu.Normalization;
 using NUnit.Framework;
 
 namespace Icu.Tests